api/requirements.txt
fastapi
pymongo
python-dateutil
uvicorn
python-dotenv
pydantic
paho-mqtt
newrelic
python-jose[cryptography]
fastapi-auth0
requests